Edit 'Write a quantum program' section
The overall flow was very clear - followed along and it worked fine (it was, in fact, the first quantum program I wrote, so mission accomplished). Since this is likely the first thing that a lot of people will read if they skip the intro material and want to get coding, I think the comments in the operations definitions should be expanded, so I added some recommendations. In addition, the routines defined here are different than the teleportation sample file in QbLibs\Samples\Teleportation - is that intentional? I can imagine that would cause confusion.