Add no-loc exclusions
This may have to do with the loc process on the live builds - once it hits a '$$', it ignores the LaTeX formatting for multiple line examples.  Look in some of the concept articles that use a lot of LaTeX, such as Oracles, DIrac notation, etc.  I think Eduardo originally figured it out and added the no-loc exclusions to the metadata in the affected articles.