Update deprecated array definition syntax
The code in this tutorial as written throws a syntax warning as follows:
```
warning QS3308: Deprecated syntax. Use [] to construct an empty array, or [x, size = n] to construct an array of x repeated n times.
```

This PR attempts to update the syntax to the new version as described in the warning. Please do let me know if another empty array definition syntax is more idiomatic. Thanks!