Update how-to-submit-quantum-include-azurecli.md
Not sure when this was changed, but referencing a quantum job (with either az quantum job show or az quantum job output) requires passing the job id with the '-id' parameter, not the '-j' parameter. Doing so with '-j' parameter will throw the error 'the following arguments are required: --job-id/-id'.